update strategy python
