+++ src/design_patterns/python/Strategy/SimpleFactoryCash.py
@@ -6,22 +6,6 @@
 @Version :   1.0
 @Contact :   pi620903@163.com
 @Desc    :   None
-@License :
-
-    (c) Copyright 2022 SheltonXiao
-
-    This program is free software: you can redistribute it and/or modify
-    it under the terms of the GNU General Public License as published by
-    the Free Software Foundation, either version 3 of the License, or(at your option) any later version.
-
-    This program is distributed in the hope that it will be useful,
-    but WITHOUT ANY WARRANTY;without even the implied warranty of
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.See the
-    GNU General Public License
-    for more details.
-
-    You should have received a copy of the GNU General Public License
-    along with this program.If not, see < https: //www.gnu.org/licenses/>.
 '''
 
 # here put the import lib
@@ -60,6 +44,7 @@ class CashReturn(CashSuper):
             result = money - money // self.__moneyCondition * self.__moneyReturn
         return result
 
+# 简单工厂方法实现
 class CashFactory(object):
     def create_cash_accept(self,typ):
         cs = CashSuper()
